+++ src/omech.cpp
@@ -7,9 +7,9 @@ GLfloat calc_e(GLfloat r, GLfloat vsq, GLfloat GM, GLfloat phi)
     // http://www.braeunig.us/space/orbmech.htm equation 4.30
     
     GLfloat t1 = r * vsq / GM - 1;
-    GLfloat t2 = SQR(t1) * SQR(cos(phi)) + SQR(sin(phi));
+    GLfloat t2 = SQR(t1) * SQR(cosf(phi)) + SQR(sinf(phi));
     
-    return sqrt(t2);
+    return sqrtf(t2);
 }
 
 GLfloat calc_a(GLfloat r, GLfloat vsq, GLfloat GM)
@@ -22,12 +22,12 @@ GLfloat calc_nu(GLfloat r, GLfloat vsq, GLfloat GM, GLfloat phi)
 {
     // http://www.braeunig.us/space/orbmech.htm equation 4.31
     GLfloat term1 = r * vsq / GM;
-    GLfloat cosPhi = cos(phi);
+    GLfloat cosPhi = cosf(phi);
     
-    GLfloat num = term1 * cosPhi * sin(phi);
+    GLfloat num = term1 * cosPhi * sinf(phi);
     GLfloat den = term1 * SQR(cosPhi) - 1.0f;
     
-    return atan2(num, den);
+    return atan2f(num, den);
 }
 
 GLfloat calc_Rp(GLfloat a, GLfloat e)
